Ensure application hard exits when unable to acquire queue connection
There were niche situations where in the event the Redis connection was
not able to be established would cause an endless execution loop that
would prevent the application from actually terminating. We mitigate this
by ensuring that the close functions will always execute the callback
function if provided in all branch scenarios.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>