Add error handler to onFailure handler
This change ensures that in the unlikely event the application is unable
to perform standard status updates that it does not create an
unhandledRejection that cascades to the main application event loop. If
the log message is reached, data integrity must be checked as there is
very likely an inconsistency between Redis and the DB that is not
resolvable during normal runtime.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>